Заменил числовые параметры константами. Убрал константы в отдельный модуль. Перенёс вспомогательные функции в модуль function.
